fix: revisão repository

+++ src/main/java/com/prim21/PRIM21/Cross/Training/repository/HorarioRepository.java
@@ -3,5 +3,5 @@ package com.prim21.PRIM21.Cross.Training.repository;
 import com.prim21.PRIM21.Cross.Training.model.Horario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HorarioRepository extends JpaRepository<Horario, Integer> {
+public interface HorarioRepository extends JpaRepository<Horario, Long> {
 }
